// src/units/enums/unit.enums.ts
export enum UnitStatus {
  AVAILABLE = 'available',
  RESERVED = 'reserved',
  SOLD = 'sold',
  NOT_FOR_SALE = 'not_for_sale',
}

export enum UnitType {
  LIVING = 'living',
  COMMERCE = 'commerce',
  PARKING = 'parking',
  STORAGE = 'storage',
}

export enum FurnishingStatus {
  WITHOUT = 'without',
  ROUGH_DRAFT = 'rough_draft',
  FINISHING = 'finishing',
  SHELL_AND_CORE = 'shell_and_core',
}